requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.1
asyncio==3.4.3
logging==0.4.9.6